Chore: deploy HTML files w/o extension
